+++ src/Data/Barbie/Bare.hs
@@ -1,55 +1,14 @@
------------------------------------------------------------------------------
--- |
--- Module      :  Data.Barbie.Bare
---
--- Sometimes one needs a type like
---  @Barbie 'Data.Functor.Identity.Identity'@ and it may feel like
--- a second-class record type, where one needs to
--- unpack values in each field. For those cases, we can leverage on
--- closed type-families:
---
--- @
--- data 'Bare'
--- data 'Covered'
---
--- type family 'Wear' t f a where
---   'Wear' 'Bare'    f a = a
---   'Wear' 'Covered' f a = f a
---
--- data SignUpForm t f
---   = SignUpForm'
---       { username  :: 'Wear' t f 'String',
---       , password  :: 'Wear' t f 'String'
---       , mailingOk :: 'Wear' t f 'Bool'
---       }
---  instance 'FunctorB' (SignUpForm 'Covered')
---  instance 'TraversableB' (SignUpForm 'Covered')
---  ...,
---  instance 'BareB' SignUpForm
---
--- type SignUpRaw  = SignUpForm 'Maybe'
--- type SignUpData = SignUpForm 'Bare'
---
--- formData = SignUpForm "jbond" "shaken007" False :: SignUpData
--- @
-
-
-----------------------------------------------------------------------------
 module Data.Barbie.Bare
+  {-# DEPRECATED "Use Barbies.Bare" #-}
   ( -- * Bare values
-    Wear
-  , Bare
-  , Covered
+    Barbies.Bare.Wear
+  , Barbies.Bare.Bare
+  , Barbies.Bare.Covered
 
     -- * Covering and stripping
-  , BareB(bstrip, bcover)
-  , bstripFrom
-  , bcoverWith
-
+  , Barbies.Bare.BareB(bstrip, bcover)
+  , Barbies.Bare.bstripFrom
+  , Barbies.Bare.bcoverWith
   ) where
 
-import Data.Barbie.Internal.Bare
-  ( Wear, Bare, Covered
-  , BareB(..)
-  , bstripFrom, bcoverWith
-  )
+import qualified Barbies.Bare