+++ ChangeLog.md
@@ -5,6 +5,7 @@
   - Remove functions deprecated on release 1.0
   - Deprecate `Data.Functor.Prod`, `(/*)` and `(/*/)`.
   - Deprecate `Data.Barbie`, in favor of `Data.Functor.Barbie`.
+  - Deprecate `Data.Barbie.Bare`, in favor of `Barbies.Bare`.
 
 ## 1.1.3.0
   - `Wear` will raise a `TypeError` instead of getting